add retry utilities
